Update main.html
Changed to fixed px boxes for now. Will need to change in the future to adjust for screen size changes. Text was not centered properly when in percentage form. Margins are now consistent. Ticker tape text still needs to be centered but will have to be turned into a rolling text later on with javascript anyway.