Update main.html
Added bootstrap library. Fixed some of the margin issues by getting rid of margins and padding on main body and html. I also replaced margin with borders on all of the elements. Added a function to clear the search box when you are clicking on it and when you hit 'esc'. Tried adding next column of charts but was unsuccessful